src/iris-mobile/requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.11.2
falcon==1.2.0
gunicorn==19.7.1
hashlib==20081119
hmac==20101005
idna==2.5
image==1.5.5
irisclient==1.2.1
Jinja2==2.9.6
json2html==1.1.1
MarkupSafe==1.0
MySQL-python==1.2.5
olefile==0.44
phonenumbers==8.5.1
Pillow==4.1.1
pypng==0.0.18
PyQRCode==1.2.1
python-gcm==0.4
python-mimeparse==1.6.0
requests==2.18.1
six==1.10.0
urllib3==1.21.1
uWSGI==2.0.15
